ai-writing-assistant-server/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
sse-starlette==2.1.3
pydantic==2.8.2
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-mock==3.12.0
tenacity
ratelimit
